chore: remove obsolete schema builder core
