+++ pkgs/standards/autoapi/autoapi/v3/schema/builder/__init__.py
@@ -2,7 +2,9 @@
 
 from .cache import _SchemaCache, _SchemaVerb
 from .extras import _merge_request_extras, _merge_response_extras
-from .core import _build_schema, _build_list_params, _strip_parent_fields
+from .build_schema import _build_schema
+from .list_params import _build_list_params
+from .strip_parent_fields import _strip_parent_fields
 
 __all__ = [
     "_build_schema",
